[써니] 2단계 로또 제출합니다. (#49)
* refactor(Lotto): 팩토리 메소드 value 타입 List로 수정

* fix(SequentialLottoNumberGenerator): numbers 인덱스 접근 에러 수정

* refactor(Lottos): Lottos는 로또들을 관리하는 책임만 갖도록 수정

* refactor(LottoTest): 중복되는 로또 생성을 위한 함수 구현

* refactor(InputView): 원시값 포장하지 않고 반환하도록 수정

* fix(RandomLottoGenerator): generateLotto List Builder 코드 수정

* feat(LottoStore): 수동, 자동 로또 생성 클래스 구현

* refactor(InputView): 수동 로또 구매 수량 입력 구현

* refactor(Rank): winningMoney 자료형 Long으로 수정

* refactor(LottoStore): 가변인자 Lotto를 받도록 수정

* refactor(World): 수동 로또 구매

* refactor(LottoCount): 자동 로또 구매 개수 계산 함수 구현

* refactor(InputView): 책임이 같은 함수 병합

* refactor(mergeLottos): LottoStore에서 구매한 로또 병합

* style(RandomLottoGenerator): map 함수 간결화

* refactor(WinStatistics): 우승 상금 합계 함수 구현

* refactor(InputView): InputView에서 입력 및 검증 후 값 반환하도록 수정

* fix(InputView): 수동 로또 0장 구매 시 수동 로또 번호 입력 문구 미출력 하도록 수정

* refactor(LottoStore): 자동 로또 구매 개수 계산 함수 구현

* refactor(LottoTest): generateLotto 함수 List로 로또 번호를 받아 생성해주도록 수정

* refactor(PurchaseMoney): 수동 로또 구매 수 입력 검증 로직 추가

* feat(LottoTicket): 수동 로또 발행 전 번호 관리 클래스 구현

* style(LottoTicket): 검증 순서 변경

* refactor(LottoStore): 수동, 자동 로또 구매 후 반환하는 로직 생성

* style(World): 무의미한 객체 생성 구문 제거

* style(PurchaseMoney): 조건문 블럭 수정

* refactor(LottoStore): 모호한 lottoCount 변수명 변경 및 반환 함수 구현

* refactor(LottoStoreTest): 테스트 코드에서만 사용되던 함수 접근 제어 수정

* fix(World): 자동 로또 개수 출력 타입 오류 수정

+++ src/main/kotlin/lotto/model/Rank.kt
@@ -1,6 +1,6 @@
 package lotto.model
 
-enum class Rank(val countOfMatch: Int, val winningMoney: Int, val matchBonus: Boolean = false) {
+enum class Rank(val countOfMatch: Int, val winningMoney: Long, val matchBonus: Boolean = false) {
     FIRST(6, 2_000_000_000),
     SECOND(5, 30_000_000, matchBonus = true),
     THIRD(5, 1_500_000),
